requirements.txt
flask
streamlit
requests
pymongo
pyyaml
google-generativeai